Extract weather icon function
