+++ src/Summary.css
@@ -10,7 +10,7 @@
     color: #88869D;
     padding-top: 80px;
     padding-bottom: 80px;
-    background-image: url("Cloud-background_grey.png");
+    background-image: url("utils/Cloud-background_blue.png");
     background-repeat: no-repeat;
     background-attachment: local;
     background-size: contain;
